+++ tests/fixtures/node/hello.js
@@ -1,0 +1,1 @@
+console.log(`HELLO ${process.argv[2] || ''} ${process.argv[3] || ''}`);